relocated Nav-Component into Subfolder and created detailedProfile Component
