+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\index.js":"1","C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\App.js":"2","C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\reportWebVitals.js":"3","C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\components\\Nav.js":"4"},{"size":515,"mtime":1611683656732,"results":"5","hashOfConfig":"6"},{"size":283,"mtime":1611682587500,"results":"7","hashOfConfig":"6"},{"size":375,"mtime":1611676375966,"results":"8","hashOfConfig":"6"},{"size":282,"mtime":1611683212974,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"4yh36o",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\index.js",[],"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\App.js",[],"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\reportWebVitals.js",[],"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\components\\Nav.js",[]]
+[{"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\index.js":"1","C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\App.js":"2","C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\reportWebVitals.js":"3","C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\components\\Nav\\index.js":"4"},{"size":515,"mtime":1611683656732,"results":"5","hashOfConfig":"6"},{"size":515,"mtime":1611684666812,"results":"7","hashOfConfig":"6"},{"size":375,"mtime":1611676375966,"results":"8","hashOfConfig":"6"},{"size":282,"mtime":1611683212974,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"4yh36o",{"filePath":"12","messages":"13","errorCount":4,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\index.js",[],"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\App.js",["18","19","20","21"],"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\reportWebVitals.js",[],"C:\\Users\\stefa\\Documents\\gitlab\\WBS\\stefano\\EricWaldemar\\githubreact\\src\\components\\Nav\\index.js",[],{"ruleId":"22","severity":2,"message":"23","line":16,"column":40,"nodeType":"24","messageId":"25","endLine":16,"endColumn":46},{"ruleId":"22","severity":2,"message":"26","line":17,"column":38,"nodeType":"24","messageId":"25","endLine":17,"endColumn":49},{"ruleId":"22","severity":2,"message":"27","line":18,"column":40,"nodeType":"24","messageId":"25","endLine":18,"endColumn":44},{"ruleId":"22","severity":2,"message":"28","line":19,"column":25,"nodeType":"24","messageId":"25","endLine":19,"endColumn":33},"no-undef","'Lookup' is not defined.","Identifier","undef","'ListWeather' is not defined.","'Home' is not defined.","'NotFound' is not defined."]
