+++ src/App.js
@@ -1,21 +1,26 @@
-import React from 'react';
-import NavBar from './components/Nav';
-
-
+import { Switch, Route } from "react-router-dom";
+import NavBar from './components/Nav/';
+ 
 function App() {
   return (
+    <>
     <div className="App">
       <h1>
         GitHub Profiles
       </h1>
-      <NavBar />
-
-        
-          
-        
-      
+      <NavBar />        
     </div>
-  );
-}
 
+
+    <Switch>
+      <Route path="/lookup" component={Lookup} />
+      <Route path="/list" component={ListWeather} />
+      <Route exact path="/" component={Home} />
+      <Route component={NotFound} />
+    </Switch>
+
+
+    </>
+  );
+} 
 export default App;
