+++ src/components/view/detailedProfile.js
@@ -1,0 +1,9 @@
+// ComponentName = DetailedProfile   
+    
+const DetailedProfile = () => { return (   
+	      <div>
+			DetailedProfile
+          </div>
+	);
+} 
+export default DetailedProfile
